Move ItemParcelable in common package
