+++ app/src/main/java/com/yesferal/hornsapp/app/util/HornsAppBottomSheetFragment.kt
@@ -2,7 +2,7 @@ package com.yesferal.hornsapp.app.util
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
-import com.yesferal.hornsapp.app.presentation.concert.ItemParcelable
+import com.yesferal.hornsapp.app.presentation.common.ItemParcelable
 import com.yesferal.hornsapp.app.presentation.concert.detail.EXTRA_PARAM_PARCELABLE
 import com.yesferal.hornsapp.app.presentation.item.ItemsFragment
